refactor mejoras a excepciones y consultas

+++ src/main/java/com/cordillera/domain/models/Mesa.java
@@ -16,6 +16,6 @@ public class Mesa implements Serializable {
     private Long idMesa;
 
     @Column(name = "numeromesa")
-    private BigDecimal numeroMesa;
+    private Integer numeroMesa;
 
 }
